Move theme declaration to app modules
