+++ app-thunderbird/src/main/kotlin/net/thunderbird/android/ThunderbirdKoinModule.kt
@@ -5,6 +5,7 @@ import app.k9mail.core.common.provider.AppNameProvider
 import app.k9mail.core.featureflag.FeatureFlagFactory
 import app.k9mail.feature.launcher.FeatureLauncherExternalContract.FeatureThemeProvider
 import app.k9mail.feature.widget.shortcut.LauncherShortcutActivity
+import app.k9mail.legacy.ui.theme.ThemeProvider
 import com.fsck.k9.AppConfig
 import com.fsck.k9.activity.MessageCompose
 import net.thunderbird.android.auth.TbOAuthConfigurationFactory
@@ -12,6 +13,7 @@ import net.thunderbird.android.dev.developmentModuleAdditions
 import net.thunderbird.android.featureflag.TbFeatureFlagFactory
 import net.thunderbird.android.provider.TbAppNameProvider
 import net.thunderbird.android.provider.TbFeatureThemeProvider
+import net.thunderbird.android.provider.TbThemeProvider
 import net.thunderbird.android.widget.appWidgetModule
 import net.thunderbird.android.widget.provider.MessageListWidgetProvider
 import net.thunderbird.android.widget.provider.UnreadWidgetProvider
@@ -27,6 +29,7 @@ val appModule = module {
     single<AppConfig> { appConfig }
     single<OAuthConfigurationFactory> { TbOAuthConfigurationFactory() }
     single<AppNameProvider> { TbAppNameProvider(androidContext()) }
+    single<ThemeProvider> { TbThemeProvider() }
     single<FeatureThemeProvider> { TbFeatureThemeProvider() }
     single<FeatureFlagFactory> { TbFeatureFlagFactory() }
