+++ legacy/ui/legacy/build.gradle.kts
@@ -8,8 +8,6 @@ dependencies {
     api(projects.legacy.ui.account)
     api(projects.legacy.ui.folder)
     api(projects.core.ui.legacy.designsystem)
-    // TODO: Remove once Theme2 is fully migrated
-    api(projects.core.ui.legacy.theme2.k9mail)
 
     implementation(projects.legacy.core)
     implementation(projects.mail.common)
