+++ legacy/ui/legacy/src/main/AndroidManifest.xml
@@ -4,7 +4,6 @@
     <uses-permission android:name="android.permission.VIBRATE" />
 
     <application
-        android:theme="@style/Theme.K9.Light"
         android:supportsRtl="true"
         >
